% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{adjust_mu}
\alias{adjust_mu}
\title{Adjust mu values based on given conditions}
\usage{
adjust_mu(mu, mu_candidate, times)
}
\arguments{
\item{mu}{Numeric, the original mu value.}

\item{mu_candidate}{Numeric, the candidate mu value for adjustment.}

\item{times}{Numeric, the threshold value for adjustment.}
}
\value{
Numeric, the adjusted mu value.
}
\description{
Adjust mu values based on given conditions
}
