% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{All.guess.update}
\alias{All.guess.update}
\title{All.guess.update: This function processes the given TEATIME object to update guesses}
\usage{
All.guess.update()
}
\value{
updated Data
}
\description{
All.guess.update: This function processes the given TEATIME object to update guesses
}
