% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{calculate_left_right_most_vaf}
\alias{calculate_left_right_most_vaf}
\title{Calculate Proportion from left, right most vaf}
\usage{
calculate_left_right_most_vaf(border_vaf, num_decimal)
}
\arguments{
\item{border_vaf}{calulated from border_vaf}

\item{num_decimal}{An integer representing the number of decimal places.}
}
\value{
A numeric value representing the Proportion.
}
\description{
Calculate Proportion from left, right most vaf
}
