% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{Compare_real_simu_peak}
\alias{Compare_real_simu_peak}
\title{Peak compare, real vs simu}
\usage{
Compare_real_simu_peak(
  peakdata,
  p,
  vaf_set,
  num_decimal,
  right_df = NULL,
  right_most.vaf = NA,
  mu_small = F,
  bac = F
)
}
\arguments{
\item{peakdata}{A dataframe contain cell.div, guess mu, etc. If have cell.div Fit case. Else, Bac case.}

\item{p}{proportion.}

\item{vaf_set}{Vaf seq used for analysis.}

\item{num_decimal}{An integer representing the number of decimal places.}

\item{right_df}{Used in Bac case.}

\item{right_most.vaf}{Used in Bac case.}

\item{mu_small}{Used in if mu should be very small (<3).}

\item{bac}{Used in Bac case.}
}
\value{
A list of comparison.
}
\description{
Peak compare, real vs simu
}
