% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{compute_AIC}
\alias{compute_AIC}
\title{Compute AIC}
\usage{
compute_AIC(log_likelihood, num_params)
}
\arguments{
\item{log_likelihood}{A numeric value representing the log likelihood of the model.}

\item{num_params}{An integer representing the number of parameters in the model.}
}
\value{
A numeric value representing the AIC.
}
\description{
This function computes the Akaike Information Criterion (AIC) for a model.
}
