% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{compute_BIC}
\alias{compute_BIC}
\title{Compute BIC}
\usage{
compute_BIC(log_likelihood, num_params, sample_size)
}
\arguments{
\item{log_likelihood}{A numeric value representing the log likelihood of the model.}

\item{num_params}{An integer representing the number of parameters in the model.}

\item{sample_size}{An integer representing the sample size.}
}
\value{
A numeric value representing the BIC.
}
\description{
This function computes the Bayesian Information Criterion (BIC) for a model.
}
