% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{determine_mupick}
\alias{determine_mupick}
\title{Determine the final mu pick based on given conditions}
\usage{
determine_mupick(fitdiff, interdiff, fitdiff2, interdiff2, fitmu, intermu)
}
\arguments{
\item{fitdiff}{Numeric, difference value for fit.}

\item{interdiff}{Numeric, difference value for inter.}

\item{fitdiff2}{Numeric, secondary difference value for fit.}

\item{interdiff2}{Numeric, secondary difference value for inter.}

\item{fitmu}{Numeric, mu value for fit.}

\item{intermu}{Numeric, mu value for inter.}
}
\value{
Numeric, the chosen mu value based on the given conditions.
}
\description{
Determine the final mu pick based on given conditions
}
