% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{evaluate_all_s}
\alias{evaluate_all_s}
\title{Evaluate all S values}
\usage{
evaluate_all_s(svalue.list, vaf.t1, p)
}
\arguments{
\item{svalue.list}{List of S values}

\item{vaf.t1}{Numeric, the VAF at t1 time}

\item{p}{Numeric, the parameter p}
}
\value{
Data frame with evaluated S values
}
\description{
Evaluate all S values
}
