% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{filter.mu.estimate}
\alias{filter.mu.estimate}
\title{Filter  mu from a range of guess}
\usage{
\method{filter}{mu.estimate}(data)
}
\arguments{
\item{data}{different cell.div and mu guess}
}
\value{
A list containing the reliable mu.
}
\description{
Filter  mu from a range of guess
}
