% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{final.process}
\alias{final.process}
\title{Process and finalize the data}
\usage{
final.process(data.rearrange, Rbest.data)
}
\arguments{
\item{data.rearrange}{Data frame}

\item{Rbest.data}{Data frame, Rbest file
This function modifies and writes the final data to a file.}
}
\value{
None
}
\description{
Process and finalize the data
}
