% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{Find.s.from.predict}
\alias{Find.s.from.predict}
\title{Find S values from prediction result}
\usage{
Find.s.from.predict(predict.result)
}
\arguments{
\item{predict.result}{Data frame with prediction results}
}
\value{
Data frame with selected S values
}
\description{
Find S values from prediction result
}
