% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{Find_p_process}
\alias{Find_p_process}
\title{This function estimates the p values, normal case.}
\usage{
Find_p_process(second.vaf, mean.a.b, upper_clonal_vaf, clonal.vaf.left)
}
\arguments{
\item{second.vaf}{cluster near clonal vafs.}

\item{mean.a.b}{peak contained in clonal vafs.}

\item{upper_clonal_vaf}{peak that is near the second cluster(mixed maybe).}

\item{clonal.vaf.left}{all vafs that belongs to upper_clonal_vaf.}
}
\value{
Guess about p with clear, meaning if clonal.vaf>1 peak, take care.
}
\description{
This function estimates the p values, normal case.
}
