% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{fit_slope_check}
\alias{fit_slope_check}
\title{Get mu guess from slope method}
\usage{
fit_slope_check(p, start.div, num_decimal, end.div = NA, vaf_set_check = NULL)
}
\arguments{
\item{p}{Proportion}

\item{start.div}{minimum cell.div}

\item{num_decimal}{An integer representing the number of decimal places.}

\item{end.div}{Set upper limit for cell.div}

\item{vaf_set_check}{provide vafs for main cluster}
}
\value{
Guess from slope method.
}
\description{
Get mu guess from slope method
}
