% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{generate_bootstrap_samples}
\alias{generate_bootstrap_samples}
\title{Generate Bootstrap Samples}
\usage{
generate_bootstrap_samples(original_data, n_samples, num_decimal)
}
\arguments{
\item{original_data}{A numeric vector of the original data.}

\item{n_samples}{An integer representing the number of bootstrap samples to generate.}

\item{num_decimal}{An integer representing the number of decimal places to round the results.}
}
\value{
A numeric vector containing the bootstrap samples.
}
\description{
This function generates bootstrap samples from the original data.
}
