% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{generate_sequence}
\alias{generate_sequence}
\title{Get a seq from S value}
\usage{
generate_sequence(n, s, set_max = F, set_max_num = 0)
}
\arguments{
\item{n}{current div}

\item{s}{s value}

\item{set_max}{if need to set max div}

\item{set_max_num}{max div}
}
\value{
Data frame with evaluated S values
}
\description{
Get a seq from S value
}
