% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{Get.S}
\alias{Get.S}
\title{Get S values}
\usage{
Get.S(p, p_thre = 1e-06)
}
\arguments{
\item{p}{Numeric, the parameter p}

\item{p_thre}{Numeric, threshold value for p}
}
\value{
Data frame with estimated S values
}
\description{
Get S values
}
