% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{Get_second_peak_ratio}
\alias{Get_second_peak_ratio}
\title{Get second peak ratio}
\usage{
Get_second_peak_ratio(simu_vaf, vaf_list, cell_list, pick.ratio = 1)
}
\arguments{
\item{simu_vaf}{Simulated VAF values}

\item{vaf_list}{List of VAF values}

\item{cell_list}{List of cell values}

\item{pick.ratio}{Numeric, the pick ratio}
}
\value{
Numeric ratio of the second peak
}
\description{
Get second peak ratio
}
