% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{get_slope}
\alias{get_slope}
\title{Prepare data for slope method}
\usage{
get_slope(df, p, result_vector, num_decimal)
}
\arguments{
\item{df}{Vaf and its cluster}

\item{p}{proportion}

\item{result_vector}{Theoretical VAF from cell.div}

\item{num_decimal}{An integer representing the number of decimal places.}
}
\value{
mu Guess.
}
\description{
Prepare data for slope method
}
