% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{Iterate_P_optimize}
\alias{Iterate_P_optimize}
\title{Given a guessed P, find mu from normal case}
\usage{
Iterate_P_optimize(
  clear,
  give.vaf,
  upper_clonal_vaf,
  clonal.vaf.left,
  second.vaf,
  num_decimal,
  p_thre = 0.05
)
}
\arguments{
\item{clear}{if we need further process}

\item{give.vaf}{Given P, coresponding Vafs}

\item{upper_clonal_vaf}{peak that is near the second cluster(mixed maybe).}

\item{clonal.vaf.left}{all vafs that belongs to upper_clonal_vaf.}

\item{second.vaf}{cluster near clonal vafs.}

\item{num_decimal}{An integer representing the number of decimal places to round the results.}

\item{p_thre}{cutoff}
}
\value{
A data frame calculated mu.
}
\description{
Given a guessed P, find mu from normal case
}
