% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{keep_longest_consecutive_rows}
\alias{keep_longest_consecutive_rows}
\title{Keep longest consecutive rows}
\usage{
keep_longest_consecutive_rows(df)
}
\arguments{
\item{df}{Data frame to filter}
}
\value{
Data frame with the longest consecutive rows
}
\description{
Keep longest consecutive rows
}
