% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{log_likelihood_mixture}
\alias{log_likelihood_mixture}
\title{Log Likelihood Mixture}
\usage{
log_likelihood_mixture(data, p_vec, depth)
}
\arguments{
\item{data}{A numeric vector of VAF values.}

\item{p_vec}{A numeric vector of probabilities.}

\item{depth}{A numeric value representing the depth.}
}
\value{
A numeric value representing the total log likelihood.
}
\description{
This function calculates the log likelihood of a mixture model for VAF data.
}
