% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{magos.114}
\alias{magos.114}
\alias{magos.139}
\alias{magos.33}
\alias{vcf.114}
\alias{vcf.139}
\alias{vcf.33}
\title{Data for MAGOS and VCF Samples}
\format{
A list containing the following data sets:
\describe{
\item{magos.114}{Magos results from sample 114, ready to run TEATIME}
\item{magos.139}{Magos results from sample 139, ready to run TEATIME}
\item{magos.33}{Magos results from sample 33, ready to run TEATIME}
\item{vcf.114}{VCF like data, REF ALT that requires MAGOS, from sample 114a}
\item{vcf.139}{VCF like data, REF ALT that requires MAGOS, from sample 139a}
\item{vcf.33}{VCF like data, REF ALT that requires MAGOS, from sample 33a}
}

An object of class \code{list} of length 5.

An object of class \code{list} of length 5.

An object of class \code{data.frame} with 1010 rows and 2 columns.

An object of class \code{data.frame} with 1106 rows and 2 columns.

An object of class \code{data.frame} with 850 rows and 2 columns.
}
\usage{
magos.114

magos.139

magos.33

vcf.114

vcf.139

vcf.33
}
\description{
Contains the data for MAGOS and VCF  used in the analysis.
}
\keyword{datasets}
