% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{maincluster.candidate.update}
\alias{maincluster.candidate.update}
\title{Main cluster vaf update if we can not find peak directly}
\usage{
maincluster.candidate.update()
}
\value{
updated Peak
}
\description{
Main cluster vaf update if we can not find peak directly
}
