% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{mu_estimation_from_peak}
\alias{mu_estimation_from_peak}
\title{Mu Estimation from Peak}
\usage{
mu_estimation_from_peak(
  vaf_set,
  p,
  num_decimal,
  p_thre,
  celldivlist = NULL,
  collect_data_check = NULL
)
}
\arguments{
\item{vaf_set}{A numeric vector of VAF values.}

\item{p}{A numeric value representing the theoretical peak.}

\item{num_decimal}{An integer representing the number of decimal places.}

\item{p_thre}{A numeric value representing the threshold for p-values.}

\item{celldivlist}{An optional numeric vector of cell division values. Default is NA, Bac case.}

\item{collect_data_check}{An optional Dataframe, we prepare it for normal case.}
}
\value{
A data frame containing the estimated mu values and related metrics.
}
\description{
This function estimates the mu values from the peak of the VAF distribution.
}
