% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{mu_estimation_small}
\alias{mu_estimation_small}
\title{This function estimates the small mu values}
\usage{
mu_estimation_small(data, p, num_decimal, p_thre)
}
\arguments{
\item{data}{A data frame containing VAF data and associated metrics.}

\item{p}{proportion.}

\item{num_decimal}{An integer representing the number of decimal places.}

\item{p_thre}{A numeric value representing the threshold for p-values.}
}
\value{
A data frame containing the estimated mu values and related metrics.
}
\description{
This function estimates the small mu values
}
