% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{mu_find_small}
\alias{mu_find_small}
\title{Find Small Mu Estimates}
\usage{
mu_find_small(vaf_set, p)
}
\arguments{
\item{vaf_set}{A numeric vector of VAF values.}

\item{p}{A numeric value representing the theoretical peak.}
}
\value{
A data frame containing the estimated mu values and related p-values.
}
\description{
Find Small Mu Estimates
}
