% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{peak_test}
\alias{peak_test}
\title{Peak compare, real vs simu}
\usage{
peak_test(df, Min_Sample_size, vaf, k, depth, num_decimal, temp_check = F)
}
\arguments{
\item{df}{A numeric vector of VAF values.}

\item{Min_Sample_size}{Check if we need bootstrap.}

\item{vaf}{right most vaf.}

\item{k}{mu values.}

\item{depth}{depth sequence.}

\item{num_decimal}{An integer representing the number of decimal places.}

\item{temp_check}{if NA is required}
}
\value{
A list of comparison.
}
\description{
Peak compare, real vs simu
}
