% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{prepare.vaf.data}
\alias{prepare.vaf.data}
\title{Prepare VAF and Depth Data}
\usage{
prepare.vaf.data(
  vafdata,
  beta,
  depth,
  magos_object,
  output.folder,
  output.prefix,
  id,
  write_final = write_final,
  debug_mode = debug_mode
)
}
\arguments{
\item{vafdata}{magos object, OR, manually, it should be dataframe with columns: vaf.1, depth.1,colors
colors refers to the cluster corresponding to each mutation}

\item{beta}{Numeric, a fixed survival rate.}

\item{depth}{Numeric, a fixed value.}

\item{magos_object}{Check if input is from magos analysis}

\item{output.folder}{folder name}

\item{output.prefix}{prefix name}

\item{id}{sample name}

\item{write_final}{Want to write final result?}

\item{debug_mode}{Want to debug?}
}
\value{
A list containing the preparation results.
}
\description{
Prepare VAF and Depth Data
}
