% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{Run.fit.maincluster}
\alias{Run.fit.maincluster}
\title{Estimate mu from main cluster, from eq in paper, Fit case}
\usage{
Run.fit.maincluster(p_thre = 0.01)
}
\arguments{
\item{p_thre}{Check if input is from magos analysis}
}
\value{
A list containing the analysis results.
}
\description{
Estimate mu from main cluster, from eq in paper, Fit case
}
