% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{Run.para.estimate.maincluster}
\alias{Run.para.estimate.maincluster}
\title{Estimate mu from main cluster, from slope method & Peak method in paper
colors refers to the cluster corresponding to each mutation}
\usage{
Run.para.estimate.maincluster(p_thre = 0.01)
}
\arguments{
\item{p_thre}{Check if input is from magos analysis}
}
\value{
A list containing the analysis results.
}
\description{
Estimate mu from main cluster, from slope method & Peak method in paper
colors refers to the cluster corresponding to each mutation
}
