% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{s.dataframe.update}
\alias{s.dataframe.update}
\title{Update dataframe for S values}
\usage{
s.dataframe.update(
  cell.list,
  p,
  vaf.t1,
  vaf_set,
  min.s.detect,
  evaluate = F,
  simulation = F
)
}
\arguments{
\item{cell.list}{List of cell values}

\item{p}{Numeric, the parameter p}

\item{vaf.t1}{Numeric, the VAF at t1 time}

\item{vaf_set}{List of VAF set values}

\item{min.s.detect}{Numeric, the minimum S detection value}

\item{evaluate}{Logical, whether to evaluate}

\item{simulation}{Logical, whether it is a simulation}
}
\value{
Data frame with updated values
}
\description{
Update dataframe for S values
}
