% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{s_update_process}
\alias{s_update_process}
\title{Update S values}
\usage{
s_update_process(
  give_n,
  cluster.result,
  vaf_set,
  min.s.detect,
  vaf.t1,
  p,
  p_thre = 1e-06
)
}
\arguments{
\item{give_n}{Integer, a given cluster}

\item{cluster.result}{Data frame with cluster results}

\item{vaf_set}{List of VAF set values}

\item{min.s.detect}{Numeric, the minimum S detection value}

\item{vaf.t1}{Numeric, the VAF at t1 time}

\item{p}{Numeric, the parameter p}

\item{p_thre}{Numeric, threshold value for p}
}
\value{
updated S value
}
\description{
Update S values
}
