% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{simu_slope}
\alias{simu_slope}
\title{Simulate Slope for VAF Set}
\usage{
simu_slope(vaf_set, p)
}
\arguments{
\item{vaf_set}{A numeric vector of VAF values.}

\item{p}{A numeric value representing the theoretical peak.}
}
\value{
A numeric value representing the simulated slope (mu_high).
}
\description{
Simulate Slope for VAF Set
}
