% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{simulate_peak}
\alias{simulate_peak}
\title{Simulate peak values}
\usage{
simulate_peak(vaf_list, cell_list, vaf_set)
}
\arguments{
\item{vaf_list}{List of VAF values}

\item{cell_list}{List of cell values}

\item{vaf_set}{List of VAF set values}
}
\value{
Numeric vector with simulated peak ratios
}
\description{
Simulate peak values
}
