% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{Simulate_ratio_peak}
\alias{Simulate_ratio_peak}
\title{Simulate peak ratio}
\usage{
Simulate_ratio_peak(p, s)
}
\arguments{
\item{p}{Numeric, the parameter p}

\item{s}{Numeric, the parameter s}
}
\value{
Numeric vector with simulated peak ratios
}
\description{
Simulate peak ratio
}
