% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{slope_simu}
\alias{slope_simu}
\title{Calculate mu}
\usage{
slope_simu(cell.div, mu, p, num_decimal)
}
\arguments{
\item{cell.div}{processed vaf data}

\item{mu}{iterated mu_est from cell.div}

\item{p}{proportion}

\item{num_decimal}{An integer representing the number of decimal places.}
}
\value{
mu list
}
\description{
Calculate mu
}
