% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATIME.r
\name{TEATIME.run}
\alias{TEATIME.run}
\title{Perform Evolutionary parameter estimation from sequencing data.}
\usage{
TEATIME.run(
  input.file,
  beta = 0.9,
  depth = NA,
  p_thre = 0.01,
  magos_object = T,
  output.folder = "./",
  output.prefix = "TEATIME",
  id = "T01",
  steps = 1:5,
  write_final = T,
  debug_mode = F,
  seed = NA
)
}
\arguments{
\item{input.file}{Three options: 1. VCF like input for magos if steps has 0; 2. magos result;
3.manually, it should be dataframe with columns: vaf.1, depth.1,colors
colors refers to the cluster corresponding to each mutation}

\item{beta}{Numeric, survival rate. Default is 0.9}

\item{depth}{Numeric,Only enter if you know exactly what the value is.}

\item{p_thre}{p value cutoff used in most function. Default is 0.01}

\item{magos_object}{boolean value Check if input is from magos result}

\item{output.folder}{the folder to write the temporary and final prediction files.}

\item{output.prefix}{prefix used to name the temporary and final prediction files.}

\item{id}{Assign unique id for each analysis}

\item{steps}{A vector of integers indicating which functions to execute. 0: run MAGOS with input; 1: prepare VAF files; 2. Rbest estimate. 3. mutation rate estimate; 4. Fitness estimate. 5 Final prediction.}

\item{write_final}{if want to write final result in .txt file}

\item{debug_mode}{If want to show error}

\item{seed}{An optional seed for reproducibility.}
}
\description{
Perform Evolutionary parameter estimation from sequencing data.
}
